% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_fragment.R
\name{dataset_fragment_to_json}
\alias{dataset_fragment_to_json}
\title{Serialize a dataset fragment to JSON}
\usage{
dataset_fragment_to_json(frag, pretty = TRUE)
}
\arguments{
\item{frag}{A \code{dataset_fragment} object.}

\item{pretty}{Pretty-print the JSON. Default \code{TRUE}.}
}
\value{
A JSON string (a \code{jsonlite::json} scalar).
}
\description{
Serialize a dataset fragment to JSON
}
